Added code to parseMetaFiles() method, now checks if there are any files ending in .qc in the directory, if there are the first such file is parsed and the filter data status is set to true.
The call to create a new plotData object now passed a QCFilterData object.

Created get/set methods for the filter status variable
Created get/set methods for the QCFilterData object
Created methods qcList() which when called checks the current filter status, if the status is true it returns the QCFilterData object, if it is false it returns null.