First cvs2git commit. Added T. Poterba changes. Rm'd CVSROOT.
This git repository has been created from the CVS repository at
http://sourceforge.net/projects/evoker/ written by Jeffrey Barrett,
James Morris et al. The cvs2svn/cvs2git tool has been used import the
CVS repostory into the present git repository.

The CVSROOT directory has been removed as it will not serve a purpose as
part of a git repository.

The following is a report of changes copied in from a bug patch and
feature extension submited by Tim Poterba.

Evoker Debugging and I/O Extension
Tim Poterba, with advice and help from Julian Maller
tpoterba@broadinstitute.org

10/24/14

Contents:
- A quick summary of what we did
- Summary of the bug
- Problems in the code
- Our fixes
- Our added I/O functionality

A quick summary of what we did
——————————————————————————————
We fixed the bug in a quick-and-dirty way with the option to do a more
thorough systematic fix later, and added a button in the settings menu
to coprint a .bim and .fam file with the .bed file, so that the output
can be easily PLINKed.

Summary of the bug
——————————————————
The stock version of Evoker runs perfectly for file input and random
access of SNPs. When dealing with single collection files (not separated
by chromosome), it outputs .bed files correctly. However, when multiple
chromosomes in a collection are altered and the .bed files are saved,
the output is unreadable. Looking closer, the files all appeared to be
exactly the same byte length – exactly as long as one of the input
.beds.

Problems in the code
————————————————————
When I tracked down the source of the issue in the code with Julian,
there were a few problems relating to the functionality allowing Evoker
to handle multiple chromosomes as part of a group:

A sample set of files to be read in is printed below for reference:

sampleset.20.bed
sampleset.20.bim
sampleset.20.bnt
sampleset.21.bed
sampleset.21.bim
sampleset.21.bnt
sampleset.fam

Evoker uses the variable collection to denote a broad group of files –
in this case, the collection is ‘sampleset’. Evoker has the ability to
handle multiple collections, and this functionality works perfectly.
Evoker also incorporates the handling of multiple chromosome files
within a collection, but there are some problems with this
implementation. Random lookup of SNPs doesn’t really work differently
when files are split by chromosome – Marker.java stores a SNP with its
byte and file location, and the hashmap in MarkerData.java storing
Markers is not split by chromosome – all the SNPs are hashed together.
While this isn’t a problem, it does lead to the root of the issue
causing all the .bed output to be the same byte size – the hashmap in
MarkerData.class called snpsPerCollection. The name itself immediately
suggests the issue – the map stores SNP totals based only on collection,
not chromosome. Later in this class, we can see a problem in the method
addFile(String bimFile, String collection, String chromosome, ...).
While this method takes both a collection and a chromosome as arguments,
it includes the following line:

snpsPerCollection.put(collection);

Only the collection is used to hash the number of SNPs in the file that
has been read, which causes a very large problem: every time a file is
added with a new chromosome in the same collection, this mapping is
overwritten. addFile is called when the file directory is opened before
any changes are made to SNP assignments, so when the user saves the .bed
files, the mapping of the current collection in snpsPerCollection points
to the number of SNPs in the last file read. This means that only one of
the BinaryDataFiles will have the correct value here, and the rest will
allocate space for an incorrect number of SNPs, creating empty data or
truncating later SNPs.

Our fixes
————————-
The way that we addressed the redundancy of the snpsPerCollection map is
by using a key that includes chromosome information. To fix this, we
changed the key assignment inMarkerData.java to a simple string
concatenation, collection+ chromosome. We added a second mapping in
MarkerData.java assigning collection to the running total of SNPs.

On the other side of the functionality, we needed to edit the pieces of
code that access this data. It is accessed in two stages – when files
are read in and when files are read in, and when new files are saved. In
BinaryData.java, we needed to replace the lookup function for numSNPs to
reflect this change. This affects the input stage, and in order to do
this we had to add a chromosome field – this means adding chromosome as
a protected variable in BinaryData.java (and therefore all of its
extensions) as well as adding it as an argument to pretty much every
constructor for BinaryData and its extensions. If there is a reason why
BinaryData.java cannot include a chromosome parameter, please let me
know!

We had to add a few more invasive changes to let the file output stage
work correctly. We had to add a DataDirectory argument to
BEDFileChanger, so that the changer could access the mapping stored in
that DataDirectory’s MarkerData. With this list of changes, we were able
to edit multiple chromosomes and output files of the correct byte size
and content.

Our added I/O functionality
———————————————————————————
Some of our colleagues noted that since Evoker outputs only a .bed file,
performing PLINK analysis on the output is more difficult since the .bim
and .fam files must be read in separately, preventing the use of the
generalized PLINK input statement.

Stock Evoker
Input:
sampleset.20.bed
sampleset.20.bim
sampleset.20.bnt
sampleset.21.bed
sampleset.21.bim
sampleset.21.bnt
sampleset.fam

Optimal Evoker
Input:
sampleset.20.bed
sampleset.20.bim
sampleset.20.bnt
sampleset.21.bed
sampleset.21.bim
sampleset.21.bnt
sampleset.fam

---Edit chromosomes 20 and 21 in Evoker, and save all BEDs---

Stock Evoker
Output:
samplesetmod.20.bed
samplesetmod.21.bed

Optimal Evoker
Output:
samplesetmod.20.bed
samplesetmod.20.bim
samplesetmod.20.fam
samplesetmod.21.bed
samplesetmod.21.bim
samplesetmod.21.fam

---Now perform PLINK on chromosome 20 SNPs---

Stock Evoker
Input commands:
--bed samplesetmod.20.bed,
--bim samplesetmod.20.bim,
--fam sampleset.fam

Optimal Evoker
Input commands:
--bfile samplesetmod.20

The optimized version has advantage of easier directory management with
UNIX, since all the files can be moved with a ‘.*’ wildcard, while in
the original output the .fam file must be moved separately. In order to
allow this, I added an option under the Settings menu in the GUI called
“Save .bed/.bim/.fam” (if a more descriptive title exists, please let me
know!). When this option is checked, a portion of code in
BEDFileChanger.java will run to copy the existing .bed and .fam files to
new names a long with each outputted .bed file. We added a lightweight
copyFile() method in BEDFileChanger.java to accomplish this.